Switch from jupyterlab-builder to jupyter-builder
