+++ .github/workflows/build.yml
@@ -18,9 +18,9 @@ env:
   # Changes here should be synchronised with pyproject.toml dependencies section.
   # If MAX installs a pre-release, may also need pre-release of jupyterlab
   MIN_LITE_VERSION: jupyterlite-core==0.7.0
-  MAX_LITE_VERSION: --pre jupyterlite-core<0.9.0 jupyterlab>=4.0.0,<5
+  MAX_LITE_VERSION: --pre jupyterlite-core<0.9.0 jupyterlab>=4.6.0,<5
 
-  LAB_VERSION: jupyterlab>=4.0.0,<5
+  LAB_VERSION: jupyterlab>=4.6.0,<5
 
 jobs:
   build:
