+++ .github/workflows/update-integration-tests.yml
@@ -28,7 +28,7 @@ jobs:
         uses: jupyterlab/maintainer-tools/.github/actions/base-setup@v1
 
       - name: Install dependencies
-        run: python -m pip install -U "jupyterlab>=4,<5"
+        run: python -m pip install -U "jupyterlab>=4.6,<5"
 
       - name: Install extension
         run: |
