+++ package.json
@@ -30,14 +30,14 @@
         "url": "https://github.com/jupyterlite/terminal.git"
     },
     "scripts": {
-        "build": "jlpm build:lib && jlpm build:labextension:dev && jlpm build:webworker",
-        "build:prod": "jlpm clean && jlpm build:lib:prod && jlpm build:labextension && jlpm build:webworker:prod",
-        "build:labextension": "jupyter labextension build .",
-        "build:labextension:dev": "jupyter labextension build --development True .",
+        "build": "jlpm build:lib && jlpm build:worker && jlpm build:labextension:dev",
+        "build:prod": "jlpm clean && jlpm build:lib:prod && jlpm build:worker:prod && jlpm build:labextension",
+        "build:labextension": "jupyter-builder build .",
+        "build:labextension:dev": "jupyter-builder build --development True .",
         "build:lib": "tsc --sourceMap",
         "build:lib:prod": "tsc",
-        "build:webworker": "webpack -c webpack.worker.config.js --env=dev",
-        "build:webworker:prod": "webpack -c webpack.worker.config.js",
+        "build:worker": "rspack --config worker.rspack.config.js --mode=development",
+        "build:worker:prod": "rspack --config worker.rspack.config.js --mode=production",
         "clean": "jlpm clean:lib",
         "clean:lib": "rimraf lib tsconfig.tsbuildinfo",
         "clean:lintcache": "rimraf .eslintcache .stylelintcache",
@@ -56,7 +56,7 @@
         "test": "jest --coverage",
         "watch": "run-p watch:src watch:labextension",
         "watch:src": "tsc -w --sourceMap",
-        "watch:labextension": "jupyter labextension watch ."
+        "watch:labextension": "jupyter-builder watch ."
     },
     "dependencies": {
         "@jupyterlab/apputils": "^4.6.0",
@@ -72,8 +72,10 @@
         "mock-socket": "^9.3.1"
     },
     "devDependencies": {
-        "@jupyterlab/builder": "^4.5.0",
+        "@jupyter/builder": "^1.0.0",
         "@jupyterlab/testutils": "^4.5.0",
+        "@rspack/cli": "^2.0.6",
+        "@rspack/core": "^2.0.6",
         "@types/jest": "^29.2.0",
         "@types/json-schema": "^7.0.11",
         "@types/react": "^18.0.26",
@@ -98,8 +100,6 @@
         "stylelint-prettier": "^4.0.0",
         "ts-loader": "^9.5.2",
         "typescript": "~5.7.0",
-        "webpack": "^5.87.0",
-        "webpack-cli": "^5.1.4",
         "yjs": "^13.5.0"
     },
     "resolutions": {
