+++ worker.rspack.config.js
@@ -1,0 +1,34 @@
+const path = require('path');
+const rules = [
+  {
+    test: /\.js$/,
+    exclude: /node_modules/,
+    loader: 'source-map-loader'
+  }
+];
+
+const resolve = {
+  fallback: {
+    fs: false,
+    child_process: false,
+    crypto: false
+  },
+  extensions: ['.js']
+};
+
+module.exports = [
+  {
+    entry: {
+      ['worker']: './lib/worker.js'
+    },
+    output: {
+      filename: '[name].js',
+      path: path.resolve(__dirname, 'lib')
+    },
+    module: {
+      rules
+    },
+    devtool: 'source-map',
+    resolve
+  }
+];
